task 4.21

+++ part4/BlogList/controlers/blogs.js
@@ -21,7 +21,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    const decodedToken = request.user
     if(!decodedToken.id) {
         return response.status(401).json({ error: 'invalid token' })
     }
@@ -29,8 +29,12 @@ blogsRouter.post('/', async (request, response) => {
     // if (!body.userId) {
     //     return response.status(400).json({ error: 'userId is required' })
     // }
-    console.log(body.userId)
+    console.log(decodedToken.id)
     const user = await User.findById(decodedToken.id)
+    if(!user) {
+        return response.status(404).json({ error: 'user not found' })
+    }
+
 
     const blog = new Blog({
         title: body.title,
@@ -46,10 +50,26 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(saveBlog)
 })
 blogsRouter.delete('/:id', async (request, response) => {
-    const blog = await Blog.findByIdAndDelete(request.params.id)
+    const decodedToken = request.user
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    console.log(blog)
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+    console.log(blog.user)
+    if (blog.user === undefined) {
+        return response.status(403).json({ error: 'blog not assigned to any user'})
+    }
+    if (blog.user && blog.user.toString() !== decodedToken.id) {
+        return response.status(403).json({ error: 'forbidden: cannot delete another user\'s blog' })
+    }
+
+    const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
-
 blogsRouter.put('/:id', async (request, response) => {
     const blog = await Blog.findByIdAndUpdate(request.params.id, request.body, { new: true })
     response.json(blog)
